Added sensitivity maps to each acquisition
Added sensitivity maps to bidsmap with suffix RB1COR. Sensitivity maps are acquired right before each acquisition (t1w, pdw, mtw). Each smap is named according to the acquisition it was acquired for. This functionality was implemented in plugin_bidsify_nk.py: When a smap is found, the next sequences are checked and the _acq- entity of the smap is defined according to the first modality of t1w, pdw, or mtw found in the next sequences.